Add equipment_status module to simplify code and reduce redundancy
