+++ part-3/sleeping_bag.rb
@@ -1,4 +1,8 @@
+require_relative "equipment_status"
+
 class SleepingBag
+  include EquipmentStatus
+
   attr_reader :style, :size, :shell
 
   def initialize(args = {})
@@ -7,31 +11,4 @@ class SleepingBag
     @shell = args.fetch(:shell) { "nylon" }
   end
 
-  def reserve
-    @reserved = true
-  end
-
-  def end_reservation
-    @reserved = false
-  end
-
-  def reserved?
-    @reserved
-  end
-
-  def available?
-    !reserved?
-  end
-
-  def damaged?
-    @damaged
-  end
-
-  def record_damage
-    @damaged = true
-  end
-
-  def repair
-    @damaged = false
-  end
 end
