+++ part-3/water_bottle.rb
@@ -1,4 +1,8 @@
+require_relative "equipment_status"
+
 class WaterBottle
+  include EquipmentStatus
+
   attr_reader :capacity, :material
 
   def initialize(args = {})
@@ -6,31 +10,4 @@ class WaterBottle
     @material = args.fetch(:material) { "tin" }
   end
 
-  def reserve
-    @reserved = true
-  end
-
-  def end_reservation
-    @reserved = false
-  end
-
-  def reserved?
-    @reserved
-  end
-
-  def available?
-    !reserved?
-  end
-
-  def damaged?
-    @damaged
-  end
-
-  def record_damage
-    @damaged = true
-  end
-
-  def repair
-    @damaged = false
-  end
 end
